Completed challenge from section 4.45.
The instructor made a nice version by passing a single object with the data to the tweet component, without having to add all the necessary fields via @Input().

I think it is a great solution and maybe a pattern to pass DTOs, Viewmodels, or any other name to pass the data into the component class.